Speed up and optimise shader build script
The rerun condition was being emitted incorrectly causing the build script to run all the time. This was probably because the game was touching the sprirv files and the build script was detecting it is a change triggering a rebuild of the shaders. The build script was simplified and the multithreading was removed due to the small amount of shaders to compile.